requirements.txt
streamlit>=1.35.0
groq>=0.9.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
requests>=2.31.0
pillow>=10.0.0
python-dotenv>=1.0.0
easyocr>=1.7.0
numpy>=1.24.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9
streamlit-mic-recorder>=0.0.4